+++ apps/admin-web/src/context/AppProvider.tsx
@@ -1,0 +1,21 @@
+import { ApolloProvider } from '@apollo/client'
+import { ThemeProvider } from '@mui/material'
+
+import Topbar from '@admin-web/common/Topbar/Topbar'
+import { defaultTheme } from '@admin-web/config/theme/index'
+import { client } from '@admin-web/services/apollo'
+
+interface AppProviderProps {
+  children: React.ReactNode
+}
+
+export function AppProvider({ children }: AppProviderProps) {
+  return (
+    <ApolloProvider client={client}>
+      <ThemeProvider theme={defaultTheme}>
+        <Topbar />
+        {children}
+      </ThemeProvider>
+    </ApolloProvider>
+  )
+}