+++ apps/admin-web/src/module/approvedReviews/index.tsx
@@ -1,0 +1,4 @@
+// Todo: get data from api
+export function ApprovedReviewsPage() {
+  return <h1>Hello</h1>
+}