+++ apps/admin-web/src/module/genEd/index.tsx
@@ -1,0 +1,11 @@
+import { Container } from '@mui/material'
+
+import GenEdList from '@admin-web/module/genEd/components/genEdList'
+
+export function GenEdPage() {
+  return (
+    <Container sx={{ mt: 4 }}>
+      <GenEdList />
+    </Container>
+  )
+}