+++ apps/api/src/clientlogging/clientlogging.module.ts
@@ -1,6 +1,7 @@
-import { Module, forwardRef, Provider } from '@nestjs/common'
+import { Module, Provider, forwardRef } from '@nestjs/common'
 import { ConfigService } from '@nestjs/config'
-import { Client } from "@opensearch-project/opensearch"
+
+import { Client } from '@opensearch-project/opensearch'
 
 import { AuthModule } from '../auth/auth.module'
 import { ClientLoggingController } from './clientlogging.controller'
@@ -8,16 +9,17 @@ import { ClientLoggingService } from './clientlogging.service'
 
 const elasticProvider: Provider<Client> = {
   provide: Client,
-  useFactory: (configService: ConfigService) => new Client({
-    node: configService.get("elasticUrl"),
-    auth: {
-      username: configService.get("elasticUsername"),
-      password: configService.get("elasticPassword"),
-    },
-    ssl: {
-      rejectUnauthorized: false,
-    }
-  }),
+  useFactory: (configService: ConfigService) =>
+    new Client({
+      node: configService.get('elasticUrl'),
+      auth: {
+        username: configService.get('elasticUsername'),
+        password: configService.get('elasticPassword'),
+      },
+      ssl: {
+        rejectUnauthorized: false,
+      },
+    }),
   inject: [ConfigService],
 }
 
@@ -27,4 +29,4 @@ const elasticProvider: Provider<Client> = {
   imports: [forwardRef(() => AuthModule)],
   exports: [ClientLoggingService],
 })
-export class ClientLoggingModule { }
+export class ClientLoggingModule {}