+++ apps/web/src/modules/CourseDetail/index.tsx
@@ -11,7 +11,6 @@ import defaultSEO from '@web/../next-seo.config'
 import { BackButton } from '@web/common/components/BackButton'
 import { useLinkBuilder } from '@web/common/hooks/useLinkBuilder'
 import { Language } from '@web/common/i18n'
-import { Review } from '@cgr/codegen'
 import { getExamDate } from '@web/common/utils/getExamDate'
 import { getExamPeriod } from '@web/common/utils/getExamPeriod'
 import { parseCourseNoFromQuery } from '@web/common/utils/parseCourseNoFromQuery'
@@ -23,6 +22,7 @@ import { ReviewProvider } from '@web/modules/CourseDetail/context/Review'
 import { generateThumbnailId } from '@web/modules/CourseThumbnailAPI/utils/generateThumbnailId'
 import { createApolloServerClient } from '@web/services/apollo'
 
+import { Review } from '@cgr/codegen'
 import {
   Course,
   GetCourseInfoDocument,