+++ apps/web/src/modules/CourseSearch/components/CourseList/index.tsx
@@ -8,9 +8,7 @@ import { Loading } from '@web/modules/CourseSearch/components/Loading'
 import { Courses } from './components/Courses'
 import { useCourseList } from './hooks'
 
-export interface CourseListProps {}
-
-export const CourseList: React.FC<CourseListProps> = () => {
+export const CourseList: React.FC = () => {
   const { courses, loading, error } = useCourseList()
 
   return (